LayoutTests/fast/frames/seamless/seamless-form-post-named-expected.txt

PASS frames[0].document.body.textContent is "DONE\n"
